+++ R/str2num.R
@@ -1,6 +1,6 @@
 #' Convert Character Numbers to Numeric
 #'
-#' If you read text containing  (possibly German, ie the decimals separated by
+#' If you read text containing  (possibly German, i.e. the decimals separated by
 #' comma and dots inserted for what they think of as readability) numbers, you
 #' may want to convert them to numeric.
 #' @param x A string representing a (possibly German) number.
@@ -13,7 +13,8 @@
 #' print(na.omit(sapply(words, str2num)), digits = 9)
 #' print(str2num(words[7]), digits = 9)
 str2num <- function(x) {
-  res <- suppressWarnings(as.numeric(sub(" .*", "", sub(",", ".", gsub("\\.", "", x)))))
+  res <- suppressWarnings(as.numeric(sub(" .*", "", sub(",", ".",
+                                                        gsub("\\.", "", x)))))
   res <- res[!is.na(res)]
   if (is_of_length_zero(res)) res <- NA
   return(res)